+++ final/db/migrate/20141202041309_create_events.rb
@@ -3,9 +3,9 @@ class CreateEvents < ActiveRecord::Migration
     create_table :events do |t|
       t.string :event_desc
       t.string :event_model
-      t.int :event_year
-      t.double :event_displacement
-      t.int :event_weight
+      t.integer :event_year
+      t.float :event_displacement
+      t.integer :event_weight
       t.references :user, index: true
       t.references :track, index: true
